requirements.txt
opencv-python-headless  # Use headless OpenCV to avoid GUI dependencies
numpy
pytesseract
streamlit
pillow
